Use Travis Caching to Speed up OpenBLAS
Travis has a caching feature, where you can "persist" a directory across
CI calls, detailed at https://docs.travis-ci.com/user/caching/

This change creates a cache in the OpenBLAS.cache/ directory (only
available on travis) which holds the OpenBLAS and netlib libraries and
headers used.  The installation script checks to see if the cache
contains the latest version of OpenBLAS by checking the commit id, and
if there is a newer commit on the OpenBLAS git repo, then the cache is
cleared and rebuilt.  There is no handling of the netlib cblas
versioning at the moment.

Each item on the built matrix gets a different cache, so we will still
build OpenBLAS at least as many times as we have go versions in the
build matrix, but after that they won't build again until OpenBLAS
updates.

+++ .travis/linux/OpenBLAS/install.sh
@@ -1,28 +1,54 @@
 set -ex
 
+CACHE_DIR=${TRAVIS_BUILD_DIR}/${BLAS_LIB}.cache
+
 # fetch fortran to build OpenBLAS
 sudo apt-get update -qq && sudo apt-get install -qq gfortran
 
-# fetch OpenBLAS
-pushd ~
-sudo git clone --depth=1 git://github.com/xianyi/OpenBLAS
+# check if cache exists
+if [ -e ${CACHE_DIR}/last_commit_id ]; then
+    echo "Cache $CACHE_DIR hit"
+    # determine current OpenBLAS master commit id and compare
+    # with commit id in cache directory
+    if [ diff <(cat ${CACHE_DIR}/last_commit_id) <(git ls-remote git://github.com/xianyi/OpenBLAS HEAD | grep -o '^\S*') ]; then
+	    echo "Cache Directory $CACHE_DIR has stale commit"
+	    # if commit is different, delete the cache
+		rm -rf ${CACHE_DIR}
+	fi
+fi
 
-# make OpenBLAS
-pushd OpenBLAS
-echo OpenBLAS $(git rev-parse HEAD)
-sudo make FC=gfortran &> /dev/null && sudo make PREFIX=/usr install
-popd
+# check if cache directory exists
+if [ ! -e ${CACHE_DIR}/last_commit_id ]; then
+    if [ -d ${CACHE_DIR} ]; then
+        # Travis automatically creates the cache directory if it does not exist,
+        # so this is needed for initialization.
+	    rm -rf ${CACHE_DIR}
+	fi
 
-# fetch cblas reference lib
-curl http://www.netlib.org/blas/blast-forum/cblas.tgz | tar -zx
+    # cache generation
+  	echo "Building cache at $CACHE_DIR"
+	mkdir ${CACHE_DIR}
+	sudo git clone --depth=1 git://github.com/xianyi/OpenBLAS
 
-# make cblas and install
-pushd CBLAS
-sudo mv Makefile.LINUX Makefile.in
-sudo BLLIB=/usr/lib/libopenblas.a make alllib
-sudo mv lib/cblas_LINUX.a /usr/lib/libcblas.a
-popd
-popd
+	pushd OpenBLAS
+	echo OpenBLAS $(git rev-parse HEAD)
+	# write commit id to cache location
+	echo $(git rev-parse HEAD) > ${CACHE_DIR}/last_commit_id
+	sudo make FC=gfortran &> /dev/null && sudo make PREFIX=${CACHE_DIR} install
+	popd
+	
+	curl http://www.netlib.org/blas/blast-forum/cblas.tgz | tar -zx
+	
+	pushd CBLAS
+	sudo mv Makefile.LINUX Makefile.in
+	sudo BLLIB=${CACHE_DIR}/lib/libopenblas.a make alllib
+	sudo mv lib/cblas_LINUX.a ${CACHE_DIR}/lib/libcblas.a
+	popd
+
+fi
+
+# copy the cache files into /usr
+sudo cp -r ${CACHE_DIR}/* /usr/
 
 # install gonum/blas against OpenBLAS
 export CGO_LDFLAGS="-L/usr/lib -lopenblas"
